refactor(web): colocate useTypeahead hook with typeahead helpers
